:recycle: Refactor home screen to main screen
