+++ lib/ui/main_screen.dart
@@ -1,6 +1,5 @@
 import "package:flutter/material.dart";
 import "package:gitdone/ui/_widgets/app_bar.dart";
-import "package:gitdone/ui/home/home_screen_view_model.dart";
 import "package:gitdone/ui/settings/settings_view.dart";
 import "package:gitdone/ui/todo_list/todo_list_view.dart";
 import "package:provider/provider.dart";
@@ -17,8 +16,8 @@ class HomeScreen extends StatefulWidget {
 class _HomeScreenState extends State<HomeScreen> {
   @override
   Widget build(final BuildContext context) => ChangeNotifierProvider(
-    create: (_) => HomeScreenViewModel(),
-    child: Consumer<HomeScreenViewModel>(
+    create: (_) => MainScreenViewModel(),
+    child: Consumer<MainScreenViewModel>(
       builder: (final context, final viewModel, final child) => Scaffold(
         appBar: const NormalAppBar(),
         body: switch (viewModel.selectedIndex) {
@@ -28,7 +27,7 @@ class _HomeScreenState extends State<HomeScreen> {
         },
         bottomNavigationBar: NavigationBar(
           destinations: const [
-            NavigationDestination(icon: Icon(Icons.inbox), label: "Home"),
+            NavigationDestination(icon: Icon(Icons.inbox), label: "Todos"),
             NavigationDestination(
               icon: Icon(Icons.settings),
               label: "Settings",
@@ -40,4 +39,18 @@ class _HomeScreenState extends State<HomeScreen> {
       ),
     ),
   );
+}
+
+/// The view model for the home screen, which manages the selected index
+class MainScreenViewModel extends ChangeNotifier {
+  int _selectedIndex = 0;
+
+  /// The currently selected index of the bottom navigation bar.
+  int get selectedIndex => _selectedIndex;
+
+  /// Updates the selected index of the bottom navigation bar.
+  void updateIndex(final int index) {
+    _selectedIndex = index;
+    notifyListeners();
+  }
 }
